Remove LoadingStrategy base class
