+++ packages/compile-solidity/compilerSupplier/loadingStrategies/index.js
@@ -1,6 +1,5 @@
 module.exports = {
   Docker: require("./Docker"),
-  LoadingStrategy: require("./LoadingStrategy"),
   Local: require("./Local"),
   Native: require("./Native"),
   VersionRange: require("./VersionRange")
